major QOL improvements - audio, damageFX, pulseFX
- add pylon and monolith pulsing
- fix all sounds in game (see note below)
- tweak player movement
- fix damage flash
- add some scenery elements

Sound was not working because I had used ScriptableObjects
to store a reference to each AudioSource that got attached
to a GameObject. This worked fine if a sound was only used
once, but if it was ever used in two or more places, only
one instance of this sound would ever play!! Not good. The
solution was to store a dictionary of sounds per script.
This should be considered a band-aid fix and works for a game
with a small scope like this, but a better solution like
FMOD makes a whole heck of a lot more sense for a centralized
audio system.